+++ Native/HookAgentDx11/Dx11PresentHook.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <windows.h>
+
+namespace ht::hook::dx11
+{
+    bool InstallPresentHook();
+    void UninstallPresentHook();
+}